chore(suite): move usePasswords hook to hooks folder
